+++ library/src/main/java/com/bumptech/glide/load/engine/prefill/BitmapPreFiller.java
@@ -55,7 +55,7 @@ public final class BitmapPreFiller {
 
   // Visible for testing.
   PreFillQueue generateAllocationOrder(PreFillType... preFillSizes) {
-    final int maxSize =
+    final long maxSize =
         memoryCache.getMaxSize() - memoryCache.getCurrentSize() + bitmapPool.getMaxSize();
 
     int totalWeight = 0;
